+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 🎉 0.7 - 04/02/2024
+- Unified single.py and multiple.py on bruteforce.py
+- Improved UI
+- Removed banner.txt file
+- Added __init__.py to directories to fix some errors
+
 # 🎉 0.6 - 31/07/2022
 - Created CHANGELOG.md file
 - Updated ascii art